requirements.txt
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==1.5.0
certifi>=2018.4.16
cffi==1.11.5
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
h5py==2.8.0
html5lib==0.9999999
imageio>=2.3.0
ipykernel==4.8.2
ipython==5.7.0
ipython-genutils==0.2.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
jupyter-tensorboard==0.1.6
kiwisolver==1.0.1
linecache2==1.0.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
mkl-fft>1.0.0
mkl-random==1.0.1
mock==2.0.0
moviepy==0.2.3.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
notebook==5.5.0
numpy==1.14.3
olefile==0.45.1
pandocfilters==1.4.2
pathlib2==2.3.2
pbr==4.0.4
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess>=0.5.2
pycparser==2.18
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyWavelets==0.5.2
pyzmq==17.0.0
scandir==1.7
scikit-image==0.13.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sk-video==1.1.10
subprocess32==3.5.2
tensorboardX==1.2
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.8.1
testpath==0.3.1
torch==0.4.0
torchvision==0.2.1
tornado==5.0.2
tqdm==4.23.4
traceback2==1.4.0
traitlets==4.3.2
unittest2==1.1.0
wcwidth==0.1.7
Werkzeug==0.14.1
